@@ -2,6 +2,7 @@
 using System.Resources;
 
 namespace REG.Core.Abstractions.Services.Exceptions;
+
 public static class ServiceExceptionExtensions
 {
     public static string LocalizedMessage(this ServiceException serviceException, ResourceManager resourceManager)
@@ -9,12 +10,14 @@ public static class ServiceExceptionExtensions
         if (string.IsNullOrEmpty(serviceException?.Message))
             return null;
 
-        if (resourceManager == null)
+        if (resourceManager is null)
             return serviceException.Message;
         try
         {
-            return serviceException.Args != null ? string.Format(resourceManager.GetString(serviceException.Message) ?? string.Empty,
-                serviceException.Args) : resourceManager.GetString(serviceException.Message);
+            return serviceException.Args is not null
+                ? string.Format(resourceManager.GetString(serviceException.Message) ?? string.Empty,
+                    serviceException.Args)
+                : resourceManager.GetString(serviceException.Message);
         }
         catch (Exception)
         {
